Simplify setup
It used to be more complex to install overmind, but it is available as a gem now.
Unfortunately, simply adding it to one of the Gemfiles will not work,
since the executable will not be found because the Gems from the Gemfiles
are installed to the vendor directory.

Also foreman is probably not needed any more as dependency - foreman is
kind of the same as overmind, but overmind has some advantages over foreman.